Added Array.cut().
Array.cut is a complex meathod for creating a new string from pieces of an
old string.  It supports begining, end, step and count parameters allowing
percice control of the result.  For convienice it wraps indecies by default
but that can be disabled for more complex queries.  Reverse iteration is
supported and many optimization shortcuts are taken, including falling back to
a single slice() call on simple queries.